requirements.txt
# Core
Django>=4.2,<5.2
djangorestframework>=3.14
django-cors-headers>=4.0

# Auth
django-rest-knox>=4.2

# Database
psycopg2-binary>=2.9
django-filter>=23.0

# API Docs
drf-spectacular>=0.26

# External APIs
requests>=2.28

# Environment
python-decouple>=3.8

# PDF Report Generation
reportlab>=4.0

# LLM — Groq (core product: AI time-projection report)
groq>=0.9

# Production server
gunicorn>=21.0

# Utilities
Pillow>=10.0
whitenoise>=6.0
